Move regions into a common file
