+++ src/plugins/three/LoadRegionPlugin.d.ts
@@ -1,9 +1,12 @@
-import { TileBoundingVolume } from '../../../three/math/TileBoundingVolume.js';
-import { Tile } from '../../../base/Tile.js';
-import { TilesRenderer } from '../../../three/TilesRenderer.js';
+import { TileBoundingVolume } from '../../../three/math/TileBoundingVolume';
+import { Tile } from '../../../base/Tile';
+import { TilesRenderer } from '../../../three/TilesRenderer';
+import { OBB } from '../../three/math/OBB';
+import { Sphere, Ray } from 'three';
 
-export interface Region {
+interface Region {
 
+	errorTarget: number;
 	intersectsTile( boundingVolume: TileBoundingVolume, tile: Tile, tilesRenderer: TilesRenderer ): boolean
 	calculateError( tile: Tile, tilesRenderer: TilesRenderer ): number
 
@@ -12,8 +15,29 @@ export interface Region {
 export class ReorientationPlugin {
 
 	addRegion( region: Region ): void;
-	removeRegion( region ): void;
-	hasRegion( region ): void;
+	removeRegion( region: Region ): void;
+	hasRegion( region: Region ): void;
 	clearRegions(): void;
+
+}
+
+export class OBBRegion extends Region {
+
+	obb: OBB;
+	constructor( errorTarget?: number, obb?: OBB );
+
+}
+
+export class RayRegion extends Regin {
+
+	ray: Ray;
+	constructor( errorTarget?: number, ray?: Ray );
+
+}
+
+export class SphereRegion {
+
+	sphere: Sphere;
+	constructor( errorTarget?: number, sphere?: Sphere );
 
 }
